Fixed compiler warnings

+++ src/zinc/hal/tiva_c/sysctl.rs
@@ -276,7 +276,7 @@ pub mod periph {
       let p = base as *const super::reg::SysCtl_rmcgc;
 
       unsafe {
-        &*p.offset(self.class as int)
+        &*p.offset(self.class as isize)
       }
     }
 
@@ -293,7 +293,7 @@ pub mod periph {
       // maybe because we also have to take the bus write time into account or
       // the CPU is more clever than I think. Anyway, looping 5 times seems to
       // work
-      for _ in range(0u, 10) {
+      for _ in range(0usize, 10) {
         unsafe {
           asm!("nop" :::: "volatile");
         }
